Requiring Updated pyOpenSSL
Requiring Updated pyOpenSSL in setup.py only, to fix pip installations on systems where OpenSSL is more up to date.  CIT cluster is still running on pyOpenSSL 0.13.1, but the OpenSSL version is old enough that this still is functional, so we will not update the requirements in the RPM or Debian packaging files yet.